requirements.txt
flask==3.0.0
oandapyV20==0.7.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
pytz==2024.1
python-binance==1.0.19